Structure, Backend: Changed directory structure, added backend test code, updated README
Split the project into two main directories: frontend and backend. At
the moment, both simply contain prototype/test code for their respective
functionalities. I've left the Makefile and README for the backend code
in the backend directory for now, though in the future it should be in
the root directory of the whole project.